+++ starsky/starsky.feature.health/Helpers/CheckForUpdatesHelper.cs
