+++ starsky/starsky.feature.health/starsky.feature.health.csproj
@@ -3,10 +3,14 @@
     <PropertyGroup>
         <TargetFramework>netcoreapp3.1</TargetFramework>
         <ProjectGuid>{d9c8e6e0-2526-4978-ad8c-b4e74993cfd8}</ProjectGuid>
-        <RootNamespace>starsky.feature.checkForUpdates</RootNamespace>
+        <RootNamespace>starsky.feature.health</RootNamespace>
         <Version>0.4.0-beta.1</Version>
     </PropertyGroup>
     <ItemGroup>
       <ProjectReference Include="..\starsky.foundation.http\starsky.foundation.http.csproj" />
+    </ItemGroup>
+    <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="3.1.8" />
+      <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks.Abstractions" Version="3.1.8" />
     </ItemGroup>
 </Project>
