+++ starsky/starskytest/starsky.feature.health/HealthCheck/DateAssemblyHealthCheckTest.cs
@@ -2,9 +2,9 @@ using System;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using starsky.Health;
+using starsky.feature.health.HealthCheck;
 
-namespace starskytest.Health
+namespace starskytest.starsky.feature.health.HealthCheck
 {
 	[TestClass]
 	public class DateAssemblyHealthCheckTest
