+++ starsky/starskytest/starsky.feature.health/HealthCheck/DiskStorageHealthCheckExtensionsTest.cs
@@ -3,10 +3,10 @@ using System.Linq;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using starsky.Health;
+using starsky.feature.health.HealthCheck;
 using starskytest.FakeCreateAn;
 
-namespace starskytest.Health
+namespace starskytest.starsky.feature.health.HealthCheck
 {
 	[TestClass]
 	public class DiskStorageHealthCheckExtensionsTest
